feat(semantics): extract statement-classification facts into @pgsql/semantics
Move facts.ts (classifyStatements + StatementFacts and friends) verbatim out of @pgsql/transform into a new @pgsql/semantics package. @pgsql/transform depends on it and re-exports the same symbols from src/facts.ts, so all existing consumers are unaffected. Transform-specific rewriting drivers stay in @pgsql/transform.
