+++ packages/semantics/src/index.ts
@@ -1,0 +1,9 @@
+export type {
+  ExtensionAction,
+  ExtensionFact,
+  QualifiedName,
+  StatementFacts,
+  StatementKind,
+  StatementSpan,
+} from './facts';
+export { classifyStatements } from './facts';
